+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Fetching pipeline test report summaries
+
+### Changed
+
+- Use pipeline test report summary build ids to set test suite/case ids
+
 ## [0.5.0] - 2024-01-26
 
 This is quite a big release that introduces a lot of breaking changes.